EES-4639 - moving ReleaseAmendmentExtensions code into ReleaseAmendmentService
